# swagger.yaml
openapi: 3.0.0
info:
  title: BiteBuddies API
  description: API documentation for BiteBuddies application services
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Account Service
  - url: http://localhost:5001
    description: Availability Service
  - url: http://localhost:5002
    description: Restaurant Service
  - url: http://localhost:5006
    description: Transcription Service
  - url: http://localhost:5007
    description: Chatbot Service
  - url: http://localhost:5010
    description: Search Service
...
paths:
...
  /account:
    post:
      tags:
        - Account
      summary: Create a new user account
      description: Register a new user in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
                - password
              properties:
                email:
                  type: string
                  description: User's email address
                name:
                  type: string
                  description: User's full name
                password:
                  type: string
                  description: User's password
                profileInfo:
                  type: string
                  description: Additional user profile information
                picture:
                  type: string
                  description: URL or Base64 of user profile picture
                given_name:
                  type: string
                  description: User's given name
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  data:
                    type: object
        "400":
          description: Bad request - account with this email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /login:
    post:
      tags:
        - Account
      summary: User login
      description: Authenticate user and get account details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - password
              properties:
                name:
                  type: string
                  description: User's name
                password:
                  type: string
                  description: User's password
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                  message:
                    type: string
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /account/all:
    get:
      tags:
        - Account
      summary: Get all accounts
      description: Retrieve a list of all user accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /account/{user_id}:
    get:
      tags:
        - Account
      summary: Get account by ID
      description: Retrieve user account details by user ID
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        "200":
          description: Account retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                  message:
                    type: string
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
    put:
      tags:
        - Account
      summary: Update account
      description: Update user account details by user ID
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                given_name:
                  type: string
                profile_info:
                  type: string
                picture:
                  type: string
      responses:
        "200":
          description: Account updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                  message:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /account/email/{email}:
    get:
      tags:
        - Account
      summary: Get account by email
      description: Retrieve user account details by email address
      parameters:
        - name: email
          in: path
          required: true
          description: Email address
          schema:
            type: string
      responses:
        "200":
          description: Account retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                  message:
                    type: string
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
    put:
      tags:
        - Account
      summary: Update account by email
      description: Update user account details by email address
      parameters:
        - name: email
          in: path
          required: true
          description: Email address
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                given_name:
                  type: string
                profile_info:
                  type: string
                picture:
                  type: string
      responses:
        "200":
          description: Account updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                  message:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /auth/linkedin:
    post:
      tags:
        - Account
      summary: LinkedIn authentication
      description: Authenticate or register a user with LinkedIn credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
              properties:
                email:
                  type: string
                  description: User's email address from LinkedIn
                name:
                  type: string
                  description: User's name from LinkedIn
                picture:
                  type: string
                  description: User's profile picture URL from LinkedIn
      responses:
        "200":
          description: LinkedIn account linked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                  message:
                    type: string
        "201":
          description: New account created with LinkedIn
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  data:
                    type: object
                  message:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string

  # =========== AVAILABILITY SERVICE ===========
...